// src/screens/home/Canvas/eventmanager/Events.type.ts
export enum CanvasType {
	static = "static",
	dynamic = "dynamic",
	both = "both",
}

/*
export type EventData =
  | {
    type: ToolsID.line;
    points?: { x: number; y: number };
    state: LineEventState;
    value?: string;
  }
  | {
    type: ToolsID.select;
    element?: ToolElements;
    state: SelectEventState;
    points?: { x: number; y: number };
  }
  | {
    type: ToolsID.text;
    points?: { x: number; y: number };
    state: TextEventState;
    value?: string;
  }
  | {
    type: ToolsID.box;
    state: CommonEventState;
  }
  | {
    type: ToolsID.chart;
    state: CommonEventState;
  }
  | {
    type: ToolsID.circle;
    state: CommonEventState;
  }
  | {
    type: ToolsID.cone;
    state: CommonEventState;
  }
  | {
    type: ToolsID.cylinder;
    state: CommonEventState;
  }
  | {
    type: ToolsID.diamond;
    state: CommonEventState;
  }
  | {
    type: ToolsID.directionalLine;
    state: CommonEventState;
  }
  | {
    type: ToolsID.eraser;
    state: CommonEventState;
  }
  | {
    type: ToolsID.highlighter;
    state: CommonEventState;
  }
  | {
    type: ToolsID.lazer;
    state: CommonEventState;
  }
  | {
    type: ToolsID.pencil;
    state: CommonEventState;
  }
  | {
    type: ToolsID.pentagon;
    state: CommonEventState;
  }
  | {
    type: ToolsID.segment;
    state: CommonEventState;
  }
  | {
    type: ToolsID.rectangle;
    state: CommonEventState;
  }
  | {
    type: ToolsID.shape;
    state: CommonEventState;
  }
  | {
    type: ToolsID.slash;
    state: CommonEventState;
  }
  | {
    type: ToolsID.table;
    state: CommonEventState;
  }
  | {
    type: DrawEvents;
    offscreen?: OffscreenCanvas;
  };

export type Point = {
  x: number;
  y: number;
};

*/